backend/requirements.txt
Flask==2.0.2
mysql-connector-python==8.0.27
Werkzeug==2.0.3
Flask-Cors==3.0.10